Backend Done

+++ BrainEX/Backend/src/controllers/chat.controller.js
@@ -1,6 +1,11 @@
-import chatModel from "../model/chat.mode.js";
+import chatModel from "../model/chat.model.js";
 import messageModel from "../model/message.model.js";
-import { generateAiResponse, generateAiTitle } from "../services/ai.service.js";
+import {
+  generateAiResponse,
+  generateAiTitle,
+  generateRandomPrompt,
+} from "../services/ai.service.js";
+import { getIO } from "../sockets/server.socket.js";
 
 /**
  * @route POST api/chat/
@@ -12,6 +17,7 @@ export async function responseGenerateController(req, res, next) {
   try {
     const { message, chat: chatId } = req.body;
     const { id } = req.user;
+    const io = getIO()
     let title = null,
       chat = null;
     if (!chatId) {
@@ -27,13 +33,20 @@ export async function responseGenerateController(req, res, next) {
       chat: chatId || chat._id,
     });
 
+    
     const messages = await messageModel
-      .find({
-        chat: chatId || chat._id,
-      })
-      .sort({ createdAt: 1 })
-      .limit(5);
+    .find({
+      chat: chatId || chat._id,
+    })
+    .sort({ createdAt: 1 })
+    .limit(5);
     const aiResponse = await generateAiResponse(messages);
+    const roomId = chatId || chat._id.toString()
+
+    for(let char of aiResponse){
+      io.to(roomId).emit("ai typing",char)
+      await new Promise(r=>setTimeout(r, 15))
+    }
 
     const aiMessage = await messageModel.create({
       role: "ai",
@@ -91,9 +104,8 @@ export async function fetchAllChats(req, res, next) {
  * @param {import('express').Request} req
  * @param {import('express').Response} res
  */
-export async function fetchMessagesOfChat(req, res,next) {
-  try{
-
+export async function fetchMessagesOfChat(req, res, next) {
+  try {
     const { chatId } = req.params;
     if (!chatId) {
       return next({
@@ -106,7 +118,7 @@ export async function fetchMessagesOfChat(req, res,next) {
       user: id,
       _id: chatId,
     });
-    
+
     if (!isValidChat) {
       return next({
         status: 403,
@@ -116,16 +128,16 @@ export async function fetchMessagesOfChat(req, res,next) {
     const messages = await messageModel.find({
       chat: chatId,
     });
-    
+
     return res.status(200).json({
       success: true,
       message: "Messages feteched successfully !",
       messages,
     });
-  }catch(err){
-    next(err)
+  } catch (err) {
+    next(err);
   }
-  }
+}
 
 /**
  * @route DELETE api/chat/:chatId/delete
@@ -133,9 +145,8 @@ export async function fetchMessagesOfChat(req, res,next) {
  * @param {import('express').Request} req
  * @param {import('express').Response} res
  */
-export async function deleteChatController(req, res,next) {
-  try{
-
+export async function deleteChatController(req, res, next) {
+  try {
     const { chatId } = req.params;
     if (!chatId) {
       return next({
@@ -143,7 +154,18 @@ export async function deleteChatController(req, res,next) {
         message: "Chat id is required !",
       });
     }
-    const chat = await chatModel.findByIdAndDelete(chatId);
+    const chat = await chatModel.findOneAndDelete({
+      _id: chatId,
+      user: req.user.id,
+    });
+
+    if (!chat) {
+      return next({
+        status: 403,
+        message: "Unauthorized or chat not found",
+      });
+    }
+
     const message = await messageModel.deleteMany({
       chat: chatId,
     });
@@ -151,22 +173,27 @@ export async function deleteChatController(req, res,next) {
       success: true,
       message: "Chat deleted successfully !",
     });
-  }catch(err){
-    next(err)
+  } catch (err) {
+    next(err);
   }
-  }
-  
-  /**
-   * @route PATCH api/chat/:chatId/updateTitle
-   * @description Updates the title of a chat
+}
+
+/**
+ * @route PATCH api/chat/:chatId/updateTitle
+ * @description Updates the title of a chat
  * @param {import('express').Request} req
  * @param {import('express').Response} res
  */
 export async function updateTitleController(req, res, next) {
-  try{
-
+  try {
     const { chatId } = req.params;
     const { title } = req.body;
+    if (!title || !title.trim()) {
+      return next({
+        status: 400,
+        message: "Title is required",
+      });
+    }
     if (!chatId) {
       return next({
         status: 400,
@@ -176,18 +203,49 @@ export async function updateTitleController(req, res, next) {
     const updatedTitle = await chatModel.findOneAndUpdate(
       {
         _id: chatId,
+        user: req.user.id,
       },
       {
         title: title,
       },
-      { new: true },
+      { returnDocument: "after" },
     );
+
+    if (!updatedTitle) {
+      return next({
+        status: 400,
+        message: "Invalid chat",
+      });
+    }
+
     return res.status(200).json({
       success: true,
       message: "Title Updated",
       updatedTitle,
     });
-  }catch(err){
-    next(err)
+  } catch (err) {
+    next(err);
   }
+}
+
+/**
+ * @route POST api/chat/random-prompt
+ * @description Generates random prompt that the user can search
+ * @param {import('express').Request} req
+ * @param {import('express').Response} res
+ */
+export async function generatePromptController(req, res) {
+  const { number } = req.body;
+  if (!number) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Number is required" });
+  }
+
+  const num = parseInt(number);
+  const response = await generateRandomPrompt(num);
+  return res.status(200).json({
+    success: true,
+    response,
+  });
 }
